Adapted model for macOS with M1 Pro chip and other improvements

+++ wan/configs/shared_config.py
@@ -7,11 +7,11 @@ wan_shared_cfg = EasyDict()
 
 # t5
 wan_shared_cfg.t5_model = 'umt5_xxl'
-wan_shared_cfg.t5_dtype = torch.bfloat16
+wan_shared_cfg.t5_dtype = torch.float32
 wan_shared_cfg.text_len = 512
 
 # transformer
-wan_shared_cfg.param_dtype = torch.bfloat16
+wan_shared_cfg.param_dtype = torch.float32
 
 # inference
 wan_shared_cfg.num_train_timesteps = 1000
