+++ wan/distributed/xdit_context_parallel.py
@@ -151,9 +151,9 @@ def usp_attn_forward(self,
                      seq_lens,
                      grid_sizes,
                      freqs,
-                     dtype=torch.bfloat16):
+                     dtype=torch.float32):
     b, s, n, d = *x.shape[:2], self.num_heads, self.head_dim
-    half_dtypes = (torch.float16, torch.bfloat16)
+    half_dtypes = (torch.float16, torch.float32)
 
     def half(x):
         return x if x.dtype in half_dtypes else x.to(dtype)
